feat: 배포 설정 수정(쿠키, nginx, cors)

+++ src/main/java/swyp_11/ssubom/global/security/jwt/JWTUtil.java
@@ -1,7 +1,6 @@
 package swyp_11.ssubom.global.security.jwt;
 
 import io.jsonwebtoken.Jwts;
-import io.jsonwebtoken.SignatureAlgorithm;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
 
@@ -19,7 +18,7 @@ public class JWTUtil {
         System.out.println("SecretKey initialized: " + secretKey);
     }
 
-    public String createJWT(String category , String kakaoId , String role , Long expiredMs) {
+    public String createJWT(String category , String kakaoId , String role , int expiredMs) {
         System.out.println("Creating JWT with category: " + category + ", kakaoId: " + kakaoId + ", role: " + role + ", expiredMs: " + expiredMs);
          String jwt = Jwts.builder()
                 .claim("category",category)
